Refactor code to handle a plugins per channel
