+++ saleor/graphql/plugins/mutations.py
@@ -5,6 +5,7 @@ from ...core.permissions import PluginsPermissions
 from ...plugins.error_codes import PluginErrorCode
 from ..core.mutations import BaseMutation
 from ..core.types.common import PluginError
+from .resolvers import resolve_plugin
 from .types import Plugin
 
 
@@ -52,7 +53,7 @@ class PluginUpdate(BaseMutation):
         channel_slug = data.get("channel")
         data = data.get("input")
         manager = info.context.plugins
-        plugin = manager.get_plugin(plugin_id)
+        plugin = manager.get_plugin(plugin_id, channel_slug)
         if not plugin:
             raise ValidationError(
                 {
@@ -61,5 +62,5 @@ class PluginUpdate(BaseMutation):
                     )
                 }
             )
-        instance = manager.save_plugin_configuration(plugin_id, channel_slug, data)
-        return PluginUpdate(plugin=instance)
+        manager.save_plugin_configuration(plugin_id, channel_slug, data)
+        return PluginUpdate(plugin=resolve_plugin(plugin_id, manager))
