+++ saleor/graphql/plugins/schema.py
@@ -1,13 +1,13 @@
 import graphene
 
 from ...core.permissions import PluginsPermissions
-from ..core.fields import BaseDjangoConnectionField
+from ..core.fields import BaseConnectionField
 from ..decorators import permission_required
 from .filters import PluginFilterInput
 from .mutations import PluginUpdate
 from .resolvers import resolve_plugin, resolve_plugins
 from .sorters import PluginSortingInput
-from .types import Plugin
+from .types import Plugin, PluginCountableConnection
 
 
 class PluginsQueries(graphene.ObjectType):
@@ -18,8 +18,9 @@ class PluginsQueries(graphene.ObjectType):
         ),
         description="Look up a plugin by ID.",
     )
-    plugins = BaseDjangoConnectionField(
-        Plugin,
+
+    plugins = BaseConnectionField(
+        PluginCountableConnection,
         filter=PluginFilterInput(description="Filtering options for plugins."),
         sort_by=PluginSortingInput(description="Sort plugins."),
         description="List of plugins.",
@@ -27,7 +28,7 @@ class PluginsQueries(graphene.ObjectType):
 
     @permission_required(PluginsPermissions.MANAGE_PLUGINS)
     def resolve_plugin(self, info, **data):
-        return resolve_plugin(info, data.get("id"), info.context.plugins)
+        return resolve_plugin(data.get("id"), info.context.plugins)
 
     @permission_required(PluginsPermissions.MANAGE_PLUGINS)
     def resolve_plugins(self, info, **kwargs):
