+++ saleor/graphql/plugins/types.py
@@ -2,36 +2,13 @@ from typing import TYPE_CHECKING, Optional
 
 import graphene
 
-from ...plugins import models
-from ...plugins.base_plugin import ConfigurationTypeField
-from ..core.connection import CountableDjangoObjectType
+from ..channel.types import Channel
+from ..core.connection import CountableConnection
 from .enums import ConfigurationTypeFieldEnum
 
 if TYPE_CHECKING:
     # flake8: noqa
     from ...plugins.base_plugin import PluginConfigurationType
-
-
-def hide_private_configuration_fields(configuration, config_structure):
-    for field in configuration:
-        name = field["name"]
-        value = field["value"]
-        if value is None:
-            continue
-        field_type = config_structure.get(name, {}).get("type")
-        if field_type == ConfigurationTypeField.PASSWORD:
-            field["value"] = "" if value else None
-
-        if field_type in [
-            ConfigurationTypeField.SECRET,
-            ConfigurationTypeField.SECRET_MULTILINE,
-        ]:
-            if not value:
-                field["value"] = None
-            elif len(value) > 4:
-                field["value"] = value[-4:]
-            else:
-                field["value"] = value[-1:]
 
 
 class ConfigurationItem(graphene.ObjectType):
@@ -45,34 +22,61 @@ class ConfigurationItem(graphene.ObjectType):
         description = "Stores information about a single configuration field."
 
 
-class Plugin(CountableDjangoObjectType):
-    id = graphene.Field(type=graphene.ID, required=True)
-    configuration = graphene.List(ConfigurationItem)
+class PluginConfiguration(graphene.ObjectType):
+    active = graphene.Boolean(
+        required=True, description="Determines if plugin is active or not."
+    )
+    channel = graphene.Field(
+        Channel,
+        description="The channel to which the plugin configuration is assigned to.",
+    )
+    configuration = graphene.List(
+        ConfigurationItem, description="Configuration of the plugin."
+    )
+
+    class Meta:
+        description = "Stores information about a configuration of plugin."
+
+
+class Plugin(graphene.ObjectType):
+    id = graphene.Field(
+        type=graphene.String, required=True, description="Identifier of the plugin."
+    )
+    name = graphene.String(description="Name of the plugin.")
+    description = graphene.String(description="Description of the plugin.")
+    global_configuration = graphene.Field(
+        PluginConfiguration,
+        description="Configuration for plugins which configuration is not channel-specific.",
+    )
+    channel_configurations = graphene.List(
+        graphene.NonNull(PluginConfiguration),
+        description="Configuration of plugins for each existing channel.",
+    )
 
     class Meta:
         description = "Plugin."
-        model = models.PluginConfiguration
-        interfaces = [graphene.relay.Node]
-        only_fields = [
-            "id",
-            "name",
-            "description",
-            "active",
-            "channel",
-            "configuration",
-        ]
-
-    def resolve_id(self: models.PluginConfiguration, _info):
-        return self.identifier
 
     @staticmethod
-    def resolve_configuration(
-        root: models.PluginConfiguration, info
-    ) -> Optional["PluginConfigurationType"]:
-        plugin = info.context.plugins.get_plugin(root.identifier)
-        if not plugin:
-            return None
-        configuration = plugin.configuration
-        if plugin.CONFIG_STRUCTURE and configuration:
-            hide_private_configuration_fields(configuration, plugin.CONFIG_STRUCTURE)
-        return configuration
+    def resolve_id(root: "Plugin", _info):
+        return root.id
+
+    @staticmethod
+    def resolve_name(root: "Plugin", _info):
+        return root.name
+
+    @staticmethod
+    def resolve_description(root: "Plugin", _info):
+        return root.description
+
+    @staticmethod
+    def resolve_global_configuration(root: "Plugin", _info):
+        return root.global_configuration
+
+    @staticmethod
+    def resolve_channel_configurations(root: "Plugin", _info):
+        return root.channel_configurations
+
+
+class PluginCountableConnection(CountableConnection):
+    class Meta:
+        node = Plugin
